// src/enums/Skill.ts
enum Skill {
  // Languages
  ts = 'TypeScript',
  dart = 'Dart',
  // Frontend Frameworks
  react = 'React',
  next = 'Next.js',
  gatsby = 'Gatsby',
  rn = 'React Native',
  flutter = 'Flutter',
  // Stylings
  tailwind = 'TailwindCSS',
  emotion = 'Emotion',
  mui = 'Material UI',
  // State Managements
  tQuery = 'TanStack Query',
  mobx = 'MobX',
  jotai = 'Jotai',
  bloc = 'BloC',
  // API / Data Communications
  graphql = 'GraphQL',
  protobuf = 'Protocol Buffer',
  // Cloud Services
  firebase = 'Firebase',
  aws = 'AWS',
  // DevOps
  fastlane = 'Fastlane',
  gradle = 'Gradle',
  gerrit = 'Gerrit',
}

export default Skill